+++ crates/ruma-signatures/src/lib.rs
@@ -53,7 +53,7 @@ pub use ruma_common::{IdParseError, SigningKeyAlgorithm};
 pub use self::{
     error::{Error, JsonError, ParseError, VerificationError},
     functions::{
-        canonical_json, check_pdu_sizes, content_hash, hash_and_sign_event, reference_hash,
+        canonical_json, check_pdu_format, content_hash, hash_and_sign_event, reference_hash,
         sign_json, verify_canonical_json_bytes, verify_event, verify_json,
     },
     keys::{Ed25519KeyPair, KeyPair, PublicKeyMap, PublicKeySet},